+++ src/lib.rs
@@ -169,6 +169,7 @@ pub fn format_unescaped(value: &Value, output: &mut String) -> Result<()> {
 pub struct TinyTemplate<'template> {
     templates: HashMap<&'template str, Template<'template>>,
     formatters: HashMap<&'template str, Box<ValueFormatter>>,
+    default_formatter: &'template ValueFormatter,
 }
 impl<'template> TinyTemplate<'template> {
     /// Create a new TinyTemplate registry. The returned registry contains no templates, and has
@@ -177,6 +178,7 @@ impl<'template> TinyTemplate<'template> {
         let mut tt = TinyTemplate {
             templates: HashMap::default(),
             formatters: HashMap::default(),
+            default_formatter: &format,
         };
         tt.add_formatter("unescaped", format_unescaped);
         tt
@@ -187,6 +189,14 @@ impl<'template> TinyTemplate<'template> {
         let template = Template::compile(text)?;
         self.templates.insert(name, template);
         Ok(())
+    }
+
+    /// Changes the default formatter from [`format`](fn.format.html) to `formatter`. Usefull in combination with [`format_unescaped`](fn.format_unescaped.html) to deactivate HTML-escaping
+    pub fn set_default_formatter<F>(&mut self, formatter: &'template F)
+    where
+        F: 'static + Fn(&Value, &mut String) -> Result<()>,
+    {
+        self.default_formatter = formatter;
     }
 
     /// Register the given formatter function under the given name.
@@ -205,7 +215,12 @@ impl<'template> TinyTemplate<'template> {
     {
         let value = serde_json::to_value(context)?;
         match self.templates.get(template) {
-            Some(tmpl) => tmpl.render(&value, &self.templates, &self.formatters),
+            Some(tmpl) => tmpl.render(
+                &value,
+                &self.templates,
+                &self.formatters,
+                self.default_formatter,
+            ),
             None => Err(Error::GenericError {
                 msg: format!("Unknown template '{}'", template),
             }),
@@ -215,5 +230,31 @@ impl<'template> TinyTemplate<'template> {
 impl<'template> Default for TinyTemplate<'template> {
     fn default() -> TinyTemplate<'template> {
         TinyTemplate::new()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[derive(Serialize)]
+    struct Context {
+        name: String,
+    }
+
+    static TEMPLATE: &'static str = "Hello {name}!";
+
+    #[test]
+    pub fn test_set_default_formatter() {
+        let mut tt = TinyTemplate::new();
+        tt.add_template("hello", TEMPLATE).unwrap();
+        tt.set_default_formatter(&format_unescaped);
+
+        let context = Context {
+            name: "<World>".to_string(),
+        };
+
+        let rendered = tt.render("hello", &context).unwrap();
+        assert_eq!(rendered, "Hello <World>!")
     }
 }